refactor(anyharness): split agent readiness and install roles
