+++ anyharness/crates/anyharness-lib/src/domains/agents/installer.rs
@@ -1,26 +1,37 @@
-use std::collections::HashMap;
 use std::path::{Path, PathBuf};
-use std::process::{Command, Stdio};
 
+use self::agent_process::{install_agent_process_artifact, is_agent_process_installable};
+use self::native::{install_native_artifact, is_native_installable};
 use super::install_lock::AgentInstallLock;
-use super::managed_npm::{
-    apply_npm_version_override, installed_npm_package_version, managed_npm_install_issue,
-    npm_package_name, npm_package_version,
-};
 use super::model::*;
-use super::readiness::resolver::{artifact_root, has_managed_registry_binary_for_names};
-use super::registry::built_in_registry;
 use super::seed;
-use crate::integrations::agent_cli::acp_registry::{self, ResolvedRegistryDistribution};
-use crate::integrations::agent_cli::executable::{
-    find_in_path, find_real_binary_in_path, is_valid_executable, make_executable,
-};
-use crate::integrations::agent_cli::launcher::{generate_launcher_script, LauncherError};
-use uuid::Uuid;
+use crate::integrations::agent_cli::launcher::LauncherError;
 
-const CURL_CONNECT_TIMEOUT: &str = "10";
-const CURL_MAX_TIME_METADATA: &str = "30";
-const CURL_MAX_TIME_DOWNLOAD: &str = "900";
+mod agent_process;
+mod downloads;
+mod native;
+mod npm;
+
+#[cfg(test)]
+use self::agent_process::{
+    install_agent_process_fallback, should_skip_existing_agent_process_install,
+};
+#[cfg(test)]
+use self::npm::{install_managed_npm_package, run_command_capture, TempDirGuard};
+#[cfg(test)]
+use crate::domains::agents::managed_npm::{
+    apply_npm_version_override, npm_package_name, npm_package_version,
+};
+#[cfg(test)]
+use crate::domains::agents::readiness::resolver::artifact_root;
+#[cfg(test)]
+use crate::integrations::agent_cli::executable::make_executable;
+#[cfg(test)]
+use std::collections::HashMap;
+#[cfg(test)]
+use std::process::Command;
+#[cfg(test)]
+use uuid::Uuid;
 
 #[derive(Debug, Clone)]
 pub struct InstalledArtifactResult {
@@ -139,1050 +150,7 @@ pub(crate) fn regenerate_seeded_agent_launchers(
     runtime_home: &Path,
     seeded_agents: &[String],
 ) -> Result<Vec<InstalledArtifactResult>, InstallError> {
-    let registry = built_in_registry();
-    let mut regenerated = Vec::new();
-    for agent in seeded_agents {
-        let Some(kind) = AgentKind::parse(agent) else {
-            continue;
-        };
-        let Some(descriptor) = registry.iter().find(|descriptor| descriptor.kind == kind) else {
-            continue;
-        };
-        if let Some(result) = regenerate_agent_process_launcher(descriptor, runtime_home)? {
-            regenerated.push(result);
-        }
-    }
-    Ok(regenerated)
-}
-
-fn is_native_installable(spec: &NativeInstallSpec) -> bool {
-    matches!(
-        spec,
-        NativeInstallSpec::DirectBinary { .. } | NativeInstallSpec::TarballRelease { .. }
-    )
-}
-
-fn is_agent_process_installable(spec: &AgentProcessInstallSpec) -> bool {
-    matches!(
-        spec,
-        AgentProcessInstallSpec::RegistryBacked { .. }
-            | AgentProcessInstallSpec::ManagedNpmPackage { .. }
-    )
-}
-
-fn regenerate_agent_process_launcher(
-    descriptor: &AgentDescriptor,
-    runtime_home: &Path,
-) -> Result<Option<InstalledArtifactResult>, InstallError> {
-    let kind = &descriptor.kind;
-    let managed_dir = artifact_root(runtime_home, kind, &ArtifactRole::AgentProcess);
-    let launcher_path = managed_dir.join(format!("{}-launcher", kind.as_str()));
-    let path_prefixes = launcher_path_prefixes(runtime_home, kind);
-    let env = managed_launcher_env(kind);
-
-    let executable_relpath = match &descriptor.agent_process.install {
-        AgentProcessInstallSpec::ManagedNpmPackage {
-            executable_relpath, ..
-        } => executable_relpath,
-        AgentProcessInstallSpec::RegistryBacked {
-            fallback:
-                AgentProcessFallback::NpmPackage {
-                    executable_relpath, ..
-                },
-            ..
-        } => executable_relpath,
-        _ => return Ok(None),
-    };
-
-    let exec_path = managed_dir.join(executable_relpath);
-    if !exec_path.exists() {
-        return Err(InstallError::MissingManagedArtifact(exec_path));
-    }
-
-    generate_launcher_script(&launcher_path, &exec_path, &[], &env, &path_prefixes)?;
-    Ok(Some(InstalledArtifactResult {
-        role: ArtifactRole::AgentProcess,
-        path: launcher_path,
-        source: "seed_launcher".into(),
-        version: None,
-    }))
-}
-
-// ---------------------------------------------------------------------------
-// Native artifact install (atomic: download to temp, rename into place)
-// ---------------------------------------------------------------------------
-
-fn install_native_artifact(
-    spec: &NativeArtifactSpec,
-    kind: &AgentKind,
-    runtime_home: &Path,
-    options: &InstallOptions,
-) -> Result<Option<InstalledArtifactResult>, InstallError> {
-    let managed_dir = artifact_root(runtime_home, kind, &ArtifactRole::NativeCli);
-    let target_path = managed_dir.join(kind.as_str());
-
-    if is_valid_executable(&target_path) && !options.reinstall {
-        return Ok(None);
-    }
-
-    if !options.reinstall {
-        if let Some(path_binary) =
-            find_in_path(kind.as_str()).filter(|path| is_valid_executable(path))
-        {
-            tracing::info!(
-                agent = kind.as_str(),
-                path = %path_binary.display(),
-                "skipping managed native install because native CLI is already available on PATH"
-            );
-            return Ok(None);
-        }
-    }
-
-    std::fs::create_dir_all(&managed_dir)?;
-    let temp_path = managed_dir.join(format!(".{}.downloading", kind.as_str()));
-
-    match &spec.install {
-        NativeInstallSpec::DirectBinary {
-            binary_url_template,
-            platform_map,
-            latest_version_url,
-        } => {
-            let platform = Platform::detect().ok_or(InstallError::UnsupportedPlatform)?;
-            let platform_str = platform_map
-                .iter()
-                .find(|(p, _)| *p == platform)
-                .map(|(_, s)| s.as_str())
-                .ok_or(InstallError::UnsupportedPlatform)?;
-
-            let version = options.native_version.clone().unwrap_or_else(|| {
-                latest_version_url
-                    .as_ref()
-                    .and_then(|url| curl_fetch_text(url).ok())
-                    .unwrap_or_else(|| "latest".into())
-                    .trim()
-                    .to_string()
-            });
-
-            let url = binary_url_template
-                .replace("{version}", &version)
-                .replace("{platform}", platform_str);
-
-            let result = curl_download_binary(&url, &temp_path);
-            if result.is_err() {
-                let _ = std::fs::remove_file(&temp_path);
-                return Err(result.unwrap_err());
-            }
-            make_executable(&temp_path)?;
-            std::fs::rename(&temp_path, &target_path)?;
-
-            Ok(Some(InstalledArtifactResult {
-                role: ArtifactRole::NativeCli,
-                path: target_path,
-                source: "managed_download".into(),
-                version: Some(version),
-            }))
-        }
-        NativeInstallSpec::TarballRelease {
-            latest_url_template,
-            versioned_url_template,
-            expected_binary_template,
-            platform_map,
-        } => {
-            let platform = Platform::detect().ok_or(InstallError::UnsupportedPlatform)?;
-            let target_triple = platform_map
-                .iter()
-                .find(|(p, _)| *p == platform)
-                .map(|(_, s)| s.as_str())
-                .ok_or(InstallError::UnsupportedPlatform)?;
-
-            let url = match &options.native_version {
-                Some(v) => versioned_url_template
-                    .replace("{version}", v)
-                    .replace("{target}", target_triple),
-                None => latest_url_template.replace("{target}", target_triple),
-            };
-            let expected_binary = expected_binary_template.replace("{target}", target_triple);
-
-            download_and_extract_tarball(&url, &expected_binary, &managed_dir, &temp_path)?;
-            make_executable(&temp_path)?;
-            std::fs::rename(&temp_path, &target_path)?;
-
-            Ok(Some(InstalledArtifactResult {
-                role: ArtifactRole::NativeCli,
-                path: target_path,
-                source: "managed_release".into(),
-                version: options.native_version.clone(),
-            }))
-        }
-        NativeInstallSpec::PathOnly { .. } | NativeInstallSpec::Manual { .. } => Ok(None),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Agent-process artifact install (registry-backed or direct managed npm)
-// ---------------------------------------------------------------------------
-
-fn install_agent_process_artifact(
-    spec: &AgentProcessArtifactSpec,
-    kind: &AgentKind,
-    runtime_home: &Path,
-    options: &InstallOptions,
-) -> Result<Option<InstalledArtifactResult>, InstallError> {
-    let managed_dir = artifact_root(runtime_home, kind, &ArtifactRole::AgentProcess);
-    let launcher_path = managed_dir.join(format!("{}-launcher", kind.as_str()));
-    let managed_native_binary =
-        artifact_root(runtime_home, kind, &ArtifactRole::NativeCli).join(kind.as_str());
-    let launcher_path_prefixes = launcher_path_prefixes(runtime_home, kind);
-    let launcher_env = managed_launcher_env(kind);
-
-    match &spec.install {
-        AgentProcessInstallSpec::RegistryBacked {
-            registry_id,
-            fallback,
-        } => {
-            if should_skip_existing_agent_process_install(
-                &spec.install,
-                kind,
-                runtime_home,
-                &launcher_path,
-                options.reinstall,
-            ) {
-                return Ok(None);
-            }
-
-            match install_from_registry(
-                registry_id,
-                kind,
-                &managed_dir,
-                &launcher_path,
-                options.agent_process_version.as_deref(),
-                &launcher_path_prefixes,
-                &launcher_env,
-            ) {
-                Ok(result) => return Ok(Some(result)),
-                Err(e) => {
-                    tracing::warn!(
-                        agent = kind.as_str(),
-                        registry_id = registry_id,
-                        error = %e,
-                        "registry install failed, falling back to local install"
-                    );
-                }
-            }
-
-            install_agent_process_fallback(
-                fallback,
-                kind,
-                &managed_dir,
-                &launcher_path,
-                options,
-                &launcher_path_prefixes,
-                &launcher_env,
-                if is_valid_executable(&managed_native_binary) {
-                    Some(managed_native_binary)
-                } else {
-                    None
-                },
-            )
-        }
-        AgentProcessInstallSpec::ManagedNpmPackage {
-            package,
-            package_subdir,
-            source_build_binary_name,
-            executable_relpath,
-        } => install_managed_npm_package(
-            package,
-            package_subdir.as_deref(),
-            source_build_binary_name.as_deref(),
-            executable_relpath,
-            &managed_dir,
-            &launcher_path,
-            options.agent_process_version.as_deref(),
-            options.reinstall,
-            &launcher_path_prefixes,
-            &launcher_env,
-            "managed_npm",
-        ),
-        AgentProcessInstallSpec::PathOnly { .. } | AgentProcessInstallSpec::Manual { .. } => {
-            Ok(None)
-        }
-    }
-}
-
-fn launcher_path_prefixes(runtime_home: &Path, kind: &AgentKind) -> Vec<PathBuf> {
-    let mut prefixes = Vec::new();
-    let managed_native_dir = artifact_root(runtime_home, kind, &ArtifactRole::NativeCli);
-    let managed_native_binary = managed_native_dir.join(kind.as_str());
-    if is_valid_executable(&managed_native_binary) {
-        prefixes.push(managed_native_dir);
-    }
-    if let Some(node_dir) = seed::bundled_node_bin_dir(runtime_home) {
-        prefixes.push(node_dir);
-    }
-    prefixes
-}
-
-fn managed_launcher_env(kind: &AgentKind) -> HashMap<String, String> {
-    let mut env = HashMap::new();
-    if *kind == AgentKind::Claude {
-        env.insert("DISABLE_AUTOUPDATER".into(), "1".into());
-    }
-    env
-}
-
-fn should_skip_existing_agent_process_install(
-    install: &AgentProcessInstallSpec,
-    kind: &AgentKind,
-    runtime_home: &Path,
-    launcher_path: &Path,
-    reinstall: bool,
-) -> bool {
-    if reinstall || !launcher_path.exists() {
-        return false;
-    }
-
-    let AgentProcessInstallSpec::RegistryBacked { fallback, .. } = install else {
-        return true;
-    };
-    let AgentProcessFallback::BinaryHint {
-        candidate_binaries, ..
-    } = fallback
-    else {
-        return true;
-    };
-
-    if !launcher_uses_binary_hint(launcher_path, candidate_binaries) {
-        return true;
-    }
-
-    has_managed_registry_binary_for_names(runtime_home, kind, candidate_binaries)
-        || candidate_binaries
-            .iter()
-            .any(|binary| find_real_binary_in_path(binary).is_some())
-}
-
-fn launcher_uses_binary_hint(launcher_path: &Path, candidate_binaries: &[String]) -> bool {
-    let Ok(contents) = std::fs::read_to_string(launcher_path) else {
-        return false;
-    };
-    candidate_binaries.iter().any(|binary| {
-        contents.contains(&format!("exec \"{binary}\""))
-            || contents.contains(&format!("exec {binary}"))
-    })
-}
-
-fn install_from_registry(
-    registry_id: &str,
-    _kind: &AgentKind,
-    managed_dir: &Path,
-    launcher_path: &Path,
-    version_override: Option<&str>,
-    path_prefixes: &[PathBuf],
-    launcher_env: &HashMap<String, String>,
-) -> Result<InstalledArtifactResult, InstallError> {
-    let resolved = acp_registry::resolve_from_registry(registry_id, version_override)
-        .map_err(|e| InstallError::RegistryFailed(e.to_string()))?;
-
-    std::fs::create_dir_all(managed_dir)?;
-
-    match resolved {
-        ResolvedRegistryDistribution::Npx {
-            package,
-            args,
-            env,
-            version,
-        } => {
-            let storage = managed_dir.join("registry_npm");
-            if storage.exists() {
-                let _ = std::fs::remove_dir_all(&storage);
-            }
-            acp_registry::install_npm_package(&storage, &package).map_err(|e| {
-                InstallError::CommandFailed {
-                    program: "npm".into(),
-                    message: e,
-                }
-            })?;
-
-            let bin_hint = registry_id.strip_suffix("-acp").unwrap_or(registry_id);
-            let bin_name = format!("{bin_hint}");
-            let npm_bin = storage.join("node_modules").join(".bin").join(&bin_name);
-            let cmd_path = if npm_bin.exists() {
-                npm_bin
-            } else {
-                find_npm_bin(&storage, registry_id)
-                    .ok_or_else(|| InstallError::MissingManagedArtifact(npm_bin))?
-            };
-
-            let env = {
-                let mut merged = env.clone();
-                merged.extend(launcher_env.clone());
-                merged
-            };
-            generate_launcher_script(launcher_path, &cmd_path, &args, &env, path_prefixes)?;
-
-            Ok(InstalledArtifactResult {
-                role: ArtifactRole::AgentProcess,
-                path: launcher_path.to_path_buf(),
-                source: "registry_npm".into(),
-                version,
-            })
-        }
-        ResolvedRegistryDistribution::Binary {
-            archive_url,
-            cmd,
-            args,
-            env,
-            version,
-        } => {
-            let storage = managed_dir.join("registry_binary");
-            let cmd_path = acp_registry::install_binary_archive(&archive_url, &cmd, &storage)
-                .map_err(|e| InstallError::FetchFailed {
-                    url: archive_url,
-                    message: e,
-                })?;
-
-            let env = {
-                let mut merged = env.clone();
-                merged.extend(launcher_env.clone());
-                merged
-            };
-            generate_launcher_script(launcher_path, &cmd_path, &args, &env, path_prefixes)?;
-
-            Ok(InstalledArtifactResult {
-                role: ArtifactRole::AgentProcess,
-                path: launcher_path.to_path_buf(),
-                source: "registry_binary".into(),
-                version,
-            })
-        }
-    }
-}
-
-fn find_npm_bin(storage_root: &Path, hint: &str) -> Option<PathBuf> {
-    let bin_dir = storage_root.join("node_modules").join(".bin");
-    if !bin_dir.exists() {
-        return None;
-    }
-    for entry in std::fs::read_dir(&bin_dir).ok()?.flatten() {
-        let name = entry.file_name();
-        let name_str = name.to_string_lossy();
-        if name_str.contains(hint) {
-            return Some(entry.path());
-        }
-    }
-    std::fs::read_dir(&bin_dir)
-        .ok()?
-        .flatten()
-        .next()
-        .map(|e| e.path())
-}
-
-fn install_managed_npm_package(
-    package: &str,
-    package_subdir: Option<&Path>,
-    source_build_binary_name: Option<&str>,
-    executable_relpath: &Path,
-    managed_dir: &Path,
-    launcher_path: &Path,
-    version_override: Option<&str>,
-    force_reinstall: bool,
-    path_prefixes: &[PathBuf],
-    launcher_env: &HashMap<String, String>,
-    source: &str,
-) -> Result<Option<InstalledArtifactResult>, InstallError> {
-    let versioned_package = apply_npm_version_override(package, version_override);
-    let exec_path = if let Some(binary_name) = source_build_binary_name {
-        managed_dir.join(platform_binary_filename(binary_name))
-    } else {
-        managed_dir.join(executable_relpath)
-    };
-    std::fs::create_dir_all(managed_dir)?;
-
-    let package_issue = if source_build_binary_name.is_none() {
-        managed_npm_install_issue(&versioned_package, managed_dir)
-    } else {
-        None
-    };
-
-    if force_reinstall || !exec_path.exists() || !launcher_path.exists() || package_issue.is_some()
-    {
-        if let Some(issue) = package_issue.as_ref() {
-            tracing::info!(
-                package = %versioned_package,
-                managed_dir = %managed_dir.display(),
-                issue = %issue,
-                "refreshing managed npm agent package"
-            );
-        }
-        tracing::info!(
-            package = %versioned_package,
-            package_subdir = ?package_subdir.map(|path| path.display().to_string()),
-            source_build_binary_name = ?source_build_binary_name,
-            managed_dir = %managed_dir.display(),
-            launcher_path = %launcher_path.display(),
-            exec_path = %exec_path.display(),
-            "installing managed npm agent package"
-        );
-        if let Some(binary_name) = source_build_binary_name {
-            install_managed_source_build_binary(&versioned_package, managed_dir, binary_name)?;
-        } else if let Some(package_subdir) = package_subdir {
-            install_managed_npm_package_from_subdir(
-                &versioned_package,
-                package_subdir,
-                managed_dir,
-            )?;
-        } else {
-            install_npm_package_into_prefix(&versioned_package, managed_dir)?;
-        }
-    }
-
-    if !exec_path.exists() {
-        tracing::error!(
-            package = %apply_npm_version_override(package, version_override),
-            managed_dir = %managed_dir.display(),
-            launcher_path = %launcher_path.display(),
-            exec_path = %exec_path.display(),
-            available_bin_entries = ?read_dir_entry_names(&managed_dir.join("node_modules").join(".bin")),
-            available_node_modules = ?read_dir_entry_names(&managed_dir.join("node_modules")),
-            "managed npm install completed but expected executable was not created"
-        );
-        return Err(InstallError::MissingManagedArtifact(exec_path));
-    }
-
-    generate_launcher_script(launcher_path, &exec_path, &[], launcher_env, path_prefixes)?;
-
-    let version = installed_npm_package_version(&versioned_package, managed_dir)
-        .or_else(|| npm_package_version(&versioned_package));
-
-    Ok(Some(InstalledArtifactResult {
-        role: ArtifactRole::AgentProcess,
-        path: launcher_path.to_path_buf(),
-        source: if source_build_binary_name.is_some() {
-            "managed_source_build".into()
-        } else {
-            source.into()
-        },
-        version,
-    }))
-}
-
-#[derive(Debug)]
-struct TempDirGuard {
-    path: PathBuf,
-}
-
-impl TempDirGuard {
-    fn new(prefix: &str) -> Result<Self, InstallError> {
-        let path = std::env::temp_dir().join(format!("anyharness-{prefix}-{}", Uuid::new_v4()));
-        std::fs::create_dir_all(&path)?;
-        Ok(Self { path })
-    }
-
-    fn path(&self) -> &Path {
-        &self.path
-    }
-}
-
-impl Drop for TempDirGuard {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_dir_all(&self.path);
-    }
-}
-
-fn install_managed_npm_package_from_subdir(
-    package: &str,
-    package_subdir: &Path,
-    managed_dir: &Path,
-) -> Result<(), InstallError> {
-    if package_subdir.is_absolute() {
-        return Err(InstallError::InvalidInstallSpec(format!(
-            "package_subdir must be relative, got {}",
-            package_subdir.display()
-        )));
-    }
-
-    let staging = TempDirGuard::new("npm-subdir")?;
-    let source_root = materialize_npm_package_source(package, staging.path())?;
-    let package_dir = resolve_npm_package_subdir(&source_root, package_subdir)?;
-    let tarball_path = pack_npm_package_dir(&package_dir, staging.path())?;
-    install_npm_package_into_prefix(&tarball_path.to_string_lossy(), managed_dir)
-}
-
-fn install_managed_source_build_binary(
-    package: &str,
-    managed_dir: &Path,
-    binary_name: &str,
-) -> Result<(), InstallError> {
-    let staging = TempDirGuard::new("source-build")?;
-    let source_root = materialize_npm_package_source(package, staging.path())?;
-    build_cargo_binary_from_source(&source_root, binary_name, managed_dir)
-}
-
-fn materialize_npm_package_source(
-    package: &str,
-    staging_root: &Path,
-) -> Result<PathBuf, InstallError> {
-    if package.starts_with("git+") {
-        return clone_git_package_source(package, staging_root);
-    }
-
-    if package.starts_with("file:") {
-        return resolve_file_package_source(package);
-    }
-
-    Err(InstallError::InvalidInstallSpec(format!(
-        "package_subdir is only supported for git+ and file: package specs, got {package}"
-    )))
-}
-
-fn clone_git_package_source(package: &str, staging_root: &Path) -> Result<PathBuf, InstallError> {
-    let without_prefix = package.strip_prefix("git+").ok_or_else(|| {
-        InstallError::InvalidInstallSpec(format!("unsupported git package spec: {package}"))
-    })?;
-    let (repo_url, git_ref) = without_prefix
-        .split_once('#')
-        .map_or((without_prefix, None), |(url, reference)| {
-            (url, Some(reference))
-        });
-    let source_root = staging_root.join("source");
-
-    run_command_capture(
-        "git",
-        Command::new("git")
-            .arg("clone")
-            .arg("--quiet")
-            .arg(repo_url)
-            .arg(&source_root),
-    )?;
-
-    if let Some(git_ref) = git_ref.filter(|reference| !reference.is_empty()) {
-        run_command_capture(
-            "git",
-            Command::new("git")
-                .arg("-C")
-                .arg(&source_root)
-                .arg("checkout")
-                .arg("--quiet")
-                .arg(git_ref),
-        )?;
-    }
-
-    Ok(source_root)
-}
-
-fn resolve_file_package_source(package: &str) -> Result<PathBuf, InstallError> {
-    if let Ok(url) = url::Url::parse(package) {
-        if url.scheme() == "file" {
-            let path = url.to_file_path().map_err(|_| {
-                InstallError::InvalidInstallSpec(format!(
-                    "could not resolve file package path: {package}"
-                ))
-            })?;
-            if path.is_dir() {
-                return Ok(path);
-            }
-            return Err(InstallError::InvalidInstallSpec(format!(
-                "file package source is not a directory: {}",
-                path.display()
-            )));
-        }
-    }
-
-    let raw_path = package.strip_prefix("file:").ok_or_else(|| {
-        InstallError::InvalidInstallSpec(format!("unsupported file package spec: {package}"))
-    })?;
-    let path = PathBuf::from(raw_path);
-    let resolved = if path.is_absolute() {
-        path
-    } else {
-        std::env::current_dir()?.join(path)
-    };
-    if resolved.is_dir() {
-        Ok(resolved)
-    } else {
-        Err(InstallError::InvalidInstallSpec(format!(
-            "file package source is not a directory: {}",
-            resolved.display()
-        )))
-    }
-}
-
-fn resolve_npm_package_subdir(
-    source_root: &Path,
-    package_subdir: &Path,
-) -> Result<PathBuf, InstallError> {
-    let source_root = source_root.canonicalize()?;
-    let package_dir = source_root.join(package_subdir);
-    if !package_dir.exists() {
-        return Err(InstallError::InvalidInstallSpec(format!(
-            "package_subdir {} does not exist inside {}",
-            package_subdir.display(),
-            source_root.display()
-        )));
-    }
-
-    let package_dir = package_dir.canonicalize()?;
-    if !package_dir.starts_with(&source_root) {
-        return Err(InstallError::InvalidInstallSpec(format!(
-            "package_subdir {} escapes source root {}",
-            package_subdir.display(),
-            source_root.display()
-        )));
-    }
-
-    let package_json = package_dir.join("package.json");
-    if !package_json.is_file() {
-        return Err(InstallError::InvalidInstallSpec(format!(
-            "package_subdir {} does not contain package.json",
-            package_subdir.display()
-        )));
-    }
-
-    Ok(package_dir)
-}
-
-fn pack_npm_package_dir(package_dir: &Path, staging_root: &Path) -> Result<PathBuf, InstallError> {
-    let output = run_command_capture(
-        "npm",
-        Command::new("npm")
-            .arg("pack")
-            .arg("--pack-destination")
-            .arg(staging_root)
-            .current_dir(package_dir),
-    )?;
-    let stdout = String::from_utf8_lossy(&output.stdout).into_owned();
-    let tarball_name = stdout
-        .lines()
-        .rev()
-        .map(str::trim)
-        .find(|line| !line.is_empty())
-        .ok_or_else(|| InstallError::CommandFailed {
-            program: "npm".into(),
-            message: format!(
-                "npm pack did not report a tarball name for {}",
-                package_dir.display()
-            ),
-        })?
-        .to_string();
-    let tarball_path = staging_root.join(&tarball_name);
-    if !tarball_path.is_file() {
-        return Err(InstallError::CommandFailed {
-            program: "npm".into(),
-            message: format!(
-                "npm pack did not create expected tarball {}",
-                tarball_path.display()
-            ),
-        });
-    }
-    Ok(tarball_path)
-}
-
-fn build_cargo_binary_from_source(
-    source_root: &Path,
-    binary_name: &str,
-    managed_dir: &Path,
-) -> Result<(), InstallError> {
-    let manifest_path = source_root.join("Cargo.toml");
-    if !manifest_path.is_file() {
-        return Err(InstallError::InvalidInstallSpec(format!(
-            "source build requested but {} does not contain Cargo.toml",
-            source_root.display()
-        )));
-    }
-
-    let target_dir = managed_dir.join("source-build-target");
-    let mut command = Command::new("cargo");
-    command
-        .arg("build")
-        .arg("-j")
-        .arg("1")
-        .arg("--bin")
-        .arg(binary_name)
-        .current_dir(source_root)
-        .env("CARGO_TARGET_DIR", &target_dir)
-        .env("CARGO_INCREMENTAL", "0")
-        .env("CARGO_BUILD_JOBS", "1")
-        .env("RUSTFLAGS", "-C debuginfo=0")
-        .stdout(Stdio::null())
-        .stderr(Stdio::piped());
-    if source_root.join("Cargo.lock").is_file() {
-        command.arg("--locked");
-    }
-    run_command_capture("cargo", &mut command)?;
-
-    let built_binary = target_dir
-        .join("debug")
-        .join(platform_binary_filename(binary_name));
-    if !built_binary.is_file() {
-        return Err(InstallError::MissingManagedArtifact(built_binary));
-    }
-
-    let installed_binary = managed_dir.join(platform_binary_filename(binary_name));
-    if let Some(parent) = installed_binary.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    std::fs::copy(&built_binary, &installed_binary)?;
-    make_executable(&installed_binary)?;
-
-    Ok(())
-}
-
-fn install_npm_package_into_prefix(package: &str, managed_dir: &Path) -> Result<(), InstallError> {
-    run_command_capture(
-        "npm",
-        Command::new("npm")
-            .args(["install", "--no-audit", "--no-fund", "--prefix"])
-            .arg(managed_dir)
-            .arg(package)
-            .stdout(Stdio::null())
-            .stderr(Stdio::piped()),
-    )
-    .map(|_| ())
-}
-
-fn platform_binary_filename(binary_name: &str) -> PathBuf {
-    if cfg!(windows) {
-        PathBuf::from(format!("{binary_name}.exe"))
-    } else {
-        PathBuf::from(binary_name)
-    }
-}
-
-fn run_command_capture(
-    program: &str,
-    command: &mut Command,
-) -> Result<std::process::Output, InstallError> {
-    let output = command.output()?;
-    if output.status.success() {
-        return Ok(output);
-    }
-
-    let status_message = output
-        .status
-        .code()
-        .map(|code| format!("exit status {code}"))
-        .unwrap_or_else(|| format!("terminated by signal: {}", output.status));
-    let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-    Err(InstallError::CommandFailed {
-        program: program.into(),
-        message: if stderr.is_empty() {
-            status_message
-        } else {
-            format!("{status_message}\n{stderr}")
-        },
-    })
-}
-
-fn read_dir_entry_names(dir: &Path) -> Vec<String> {
-    let mut entries: Vec<String> = std::fs::read_dir(dir)
-        .ok()
-        .into_iter()
-        .flat_map(|entries| entries.flatten())
-        .filter_map(|entry| entry.file_name().into_string().ok())
-        .collect();
-    entries.sort();
-    entries
-}
-
-fn install_agent_process_fallback(
-    fallback: &AgentProcessFallback,
-    kind: &AgentKind,
-    managed_dir: &Path,
-    launcher_path: &Path,
-    options: &InstallOptions,
-    path_prefixes: &[PathBuf],
-    launcher_env: &HashMap<String, String>,
-    managed_native_binary: Option<PathBuf>,
-) -> Result<Option<InstalledArtifactResult>, InstallError> {
-    match fallback {
-        AgentProcessFallback::NpmPackage {
-            package,
-            package_subdir,
-            source_build_binary_name,
-            executable_relpath,
-        } => install_managed_npm_package(
-            package,
-            package_subdir.as_deref(),
-            source_build_binary_name.as_deref(),
-            executable_relpath,
-            managed_dir,
-            launcher_path,
-            options.agent_process_version.as_deref(),
-            options.reinstall,
-            path_prefixes,
-            launcher_env,
-            "fallback_npm",
-        ),
-        AgentProcessFallback::NativeSubcommand { args } => {
-            std::fs::create_dir_all(managed_dir)?;
-            let native_exec = managed_native_binary.unwrap_or_else(|| PathBuf::from(kind.as_str()));
-            generate_launcher_script(
-                launcher_path,
-                &native_exec,
-                args,
-                launcher_env,
-                path_prefixes,
-            )?;
-
-            Ok(Some(InstalledArtifactResult {
-                role: ArtifactRole::AgentProcess,
-                path: launcher_path.to_path_buf(),
-                source: "native_subcommand".into(),
-                version: None,
-            }))
-        }
-        AgentProcessFallback::BinaryHint {
-            candidate_binaries,
-            args,
-        } => {
-            let Some(bin) = candidate_binaries
-                .iter()
-                .find_map(|binary| find_real_binary_in_path(binary))
-            else {
-                return Err(InstallError::NotInstallable);
-            };
-            std::fs::create_dir_all(managed_dir)?;
-            generate_launcher_script(launcher_path, &bin, args, launcher_env, path_prefixes)?;
-
-            Ok(Some(InstalledArtifactResult {
-                role: ArtifactRole::AgentProcess,
-                path: launcher_path.to_path_buf(),
-                source: "binary_hint".into(),
-                version: None,
-            }))
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Download / extraction helpers (with timeouts)
-// ---------------------------------------------------------------------------
-
-fn curl_fetch_text(url: &str) -> Result<String, InstallError> {
-    let output = Command::new("curl")
-        .args([
-            "-fsSL",
-            "--connect-timeout",
-            CURL_CONNECT_TIMEOUT,
-            "--max-time",
-            CURL_MAX_TIME_METADATA,
-            url,
-        ])
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped())
-        .output()
-        .map_err(|e| InstallError::FetchFailed {
-            url: url.into(),
-            message: e.to_string(),
-        })?;
-
-    if !output.status.success() {
-        return Err(InstallError::FetchFailed {
-            url: url.into(),
-            message: String::from_utf8_lossy(&output.stderr).trim().to_string(),
-        });
-    }
-
-    Ok(String::from_utf8_lossy(&output.stdout).to_string())
-}
-
-fn curl_download_binary(url: &str, dest: &Path) -> Result<(), InstallError> {
-    let output = Command::new("curl")
-        .args([
-            "-fsSL",
-            "--connect-timeout",
-            CURL_CONNECT_TIMEOUT,
-            "--max-time",
-            CURL_MAX_TIME_DOWNLOAD,
-            "-o",
-        ])
-        .arg(dest)
-        .arg(url)
-        .stdout(Stdio::null())
-        .stderr(Stdio::piped())
-        .output()
-        .map_err(|e| InstallError::FetchFailed {
-            url: url.into(),
-            message: e.to_string(),
-        })?;
-
-    if !output.status.success() {
-        return Err(InstallError::FetchFailed {
-            url: url.into(),
-            message: String::from_utf8_lossy(&output.stderr).trim().to_string(),
-        });
-    }
-
-    Ok(())
-}
-
-fn download_and_extract_tarball(
-    url: &str,
-    expected_binary: &str,
-    managed_dir: &Path,
-    target_path: &Path,
-) -> Result<(), InstallError> {
-    let staging_dir = managed_dir.join("_staging");
-    let _ = std::fs::remove_dir_all(&staging_dir);
-    std::fs::create_dir_all(&staging_dir)?;
-
-    let output = Command::new("sh")
-        .args([
-            "-c",
-            &format!(
-                "curl -fsSL --connect-timeout {} --max-time {} '{}' | tar xz -C '{}'",
-                CURL_CONNECT_TIMEOUT,
-                CURL_MAX_TIME_DOWNLOAD,
-                url,
-                staging_dir.display()
-            ),
-        ])
-        .stdout(Stdio::null())
-        .stderr(Stdio::piped())
-        .output()
-        .map_err(|e| InstallError::FetchFailed {
-            url: url.into(),
-            message: e.to_string(),
-        })?;
-
-    if !output.status.success() {
-        let _ = std::fs::remove_dir_all(&staging_dir);
-        return Err(InstallError::FetchFailed {
-            url: url.into(),
-            message: String::from_utf8_lossy(&output.stderr).trim().to_string(),
-        });
-    }
-
-    let extracted = staging_dir.join(expected_binary);
-    if extracted.exists() {
-        std::fs::rename(&extracted, target_path)?;
-    } else if let Some(found) = find_binary_in_dir(&staging_dir, expected_binary) {
-        std::fs::rename(&found, target_path)?;
-    } else {
-        let _ = std::fs::remove_dir_all(&staging_dir);
-        return Err(InstallError::MissingManagedArtifact(PathBuf::from(
-            expected_binary,
-        )));
-    }
-
-    let _ = std::fs::remove_dir_all(&staging_dir);
-    Ok(())
-}
-
-fn find_binary_in_dir(dir: &Path, name: &str) -> Option<PathBuf> {
-    let entries = std::fs::read_dir(dir).ok()?;
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.is_file() && path.file_name().and_then(|n| n.to_str()) == Some(name) {
-            return Some(path);
-        }
-        if path.is_dir() {
-            if let Some(found) = find_binary_in_dir(&path, name) {
-                return Some(found);
-            }
-        }
-    }
-    None
+    agent_process::regenerate_seeded_agent_launchers(runtime_home, seeded_agents)
 }
 
 #[cfg(test)]
