+++ anyharness/crates/anyharness-lib/src/domains/agents/readiness/mod.rs
@@ -1,2 +1,7 @@
+mod artifacts;
+mod compatibility;
 pub mod launch_options;
+mod overrides;
+mod paths;
 pub mod resolver;
+mod status;
