+++ scripts/max_lines_allowlist.txt
@@ -15,8 +15,6 @@ anyharness/crates/anyharness-lib/src/api/http/sessions.rs 1611 existing AnyHarne
 anyharness/crates/anyharness-lib/src/api/http/workspaces.rs 1482 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/api/openapi.rs 854 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/api/router_tests.rs 1216 existing AnyHarness oversized test file
-anyharness/crates/anyharness-lib/src/domains/agents/installer.rs 1632 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/domains/agents/readiness/resolver.rs 1125 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/domains/cowork/runtime.rs 1882 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/domains/mobility/service.rs 1534 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/domains/plans/service.rs 653 existing AnyHarness oversized file
